refactor(record-modal): move to /record/ folder
