+++ src/components/custom-drawer/drawer-navigator/index.js
@@ -16,7 +16,7 @@ import MainConferenceScreen from '../../../screens/main-conference-screen';
 import SelectLanguageScreen from '../../../screens/select-language-screen';
 import InsideBreakoutRoomScreen from '../../../screens/inside-breakout-room-screen';
 import FullscreenWrapperScreen from '../../../screens/fullscreen-wrapper-screen';
-import RecordingIndicator from '../../recording-indicator';
+import RecordingIndicator from '../../record/record-indicator';
 import Colors from '../../../constants/colors';
 import Styled from './styles';
 import usePrevious from '../../../hooks/use-previous';
