+++ src/components/modal/index.js
@@ -2,8 +2,8 @@ import React from 'react';
 import { useSelector } from 'react-redux';
 import BreakoutInviteModal from '../../screens/breakout-room-screen/breakout-invite-modal';
 import CantCreatePollModal from '../../screens/poll-screen/modals/cant-create-poll';
-import RecordControlsModal from './record-controls-modal';
-import RecordStatusModal from './record-status-modal';
+import RecordControlsModal from '../record/modals/record-controls-modal';
+import RecordStatusModal from '../record/modals/record-status-modal';
 import ReceivePollModal from '../../screens/poll-screen/modals/receive-poll';
 import PublishedPollModal from '../../screens/poll-screen/modals/published-poll';
 import AudioDeviceSelectorModal from '../actions-bar/audio-device-selector-control/audio-device-selector-modal';
