+++ src/components/record/modals/record-controls-modal/index.js
@@ -4,9 +4,9 @@ import { Pressable } from 'react-native';
 import { Modal } from 'react-native-paper';
 import { useDispatch, useSelector } from 'react-redux';
 import { useTranslation } from 'react-i18next';
-import { hide } from '../../../store/redux/slices/wide-app/modal';
-import { selectRecordMeeting } from '../../../store/redux/slices/record-meetings';
-import Colors from '../../../constants/colors';
+import { hide } from '../../../../store/redux/slices/wide-app/modal';
+import { selectRecordMeeting } from '../../../../store/redux/slices/record-meetings';
+import Colors from '../../../../constants/colors';
 import ViewerService from '../record-status-modal/service';
 import Service from './service';
 import Styled from './styles';
