+++ src/components/record/modals/record-status-modal/styles.js
@@ -1,7 +1,7 @@
 import styled from 'styled-components/native';
 import { View, Text } from 'react-native';
 import Icon from '@expo/vector-icons/MaterialIcons';
-import Colors from '../../../constants/colors';
+import Colors from '../../../../constants/colors';
 
 const ModalContainer = styled(View)`
   justify-content: center;
